core/requirements.txt
Khayyam
django==4.2.8
django-decouple==2.1
djangorestframework==3.14.0
filelock==3.13.1
markdown==3.5.1
pillow==10.1.0
django-storages
boto3
python-dotenv
django-filter
django-ckeditor
celery
django-multiupload
django-jsonfield  
drf-yasg
requests
markdown       
django-filter  
django-cors-headers
django-debug-toolbar
psycopg2-binary
googletrans==4.0.0-rc1
bs4
pytz
django-admin-interface
django_jalali
WeasyPrint

